Add def of "discard"
Where to describe discards in the spec? I don’t think it warrants being in the “Basic concepts” chapter. The “Variables” chapter looks like the best fit. My initial thought was to make discards a new variable category. However, as a discard is a pseudo-variable, my current thinking is to describe it in a new top-level section in the variables chapter. Basically, we don’t want to have to change any existing prose w.r.t variables to exclude discards. Better to make them sort-of-but-not-quite-variables! And while a variable name is a *variable_reference*, a discard is not.